use mbox size in loading bar
The loading bar now uses bytes-processed instead of emails-processed as
its unit. The problem with emails-processed is that we didn't know how
many emails were in the file before reading it, so the total was
hard-coded. OTOH, we can of course get the file's size before reading.

Loading bar now looks like:

   2.06G/2.17G [00:56<00:06, 16.1MB/s]